SAGE-1246: fix issue where docker registries won't start in airgap on first boot
The default RestartSec is set to 100ms which causes the service to restart
too quickly for the allowed "burst restart" policy. This causes
systemd to _not_ retry restarting the service anymore. Relaxing the
restart timer to 30s allows the restarts to _not_ violate the
"burst limit" and allows for indefinite restarts in air-gap conditions.